Move ip_get_ti to input.c
