+++ input.c
@@ -32,6 +32,7 @@
 #include "misc.h"
 #include "debug.h"
 #include "ui_curses.h"
+#include "track_info.h"
 #include "locking.h"
 #ifdef HAVE_CONFIG
 #include "config/libdir.h"
@@ -1047,4 +1048,32 @@ void ip_dump_plugins(void)
 		printf("\n");
 	}
 	ip_unlock();
+}
+
+struct track_info *ip_get_ti(const char *filename)
+{
+	struct track_info *ti = NULL;
+	struct input_plugin *ip;
+	struct keyval *comments;
+	int rc;
+
+	ip = ip_new(filename);
+	rc = ip_open(ip);
+	if (rc) {
+		ip_delete(ip);
+		return NULL;
+	}
+
+	rc = ip_read_comments(ip, &comments);
+	if (!rc) {
+		ti = track_info_new(filename);
+		track_info_set_comments(ti, comments);
+		ti->duration = ip_duration(ip);
+		ti->bitrate = ip_bitrate(ip);
+		ti->codec = ip_codec(ip);
+		ti->codec_profile = ip_codec_profile(ip);
+		ti->mtime = ip_is_remote(ip) ? -1 : file_get_mtime(filename);
+	}
+	ip_delete(ip);
+	return ti;
 }
